fixed error haha

+++ index.js
@@ -89,9 +89,11 @@ client.on("message", async message => {
     var bannedword = ["faggot","nigger"]
    
 
-    bannedword.forEach(element =>  if (message.content.includes(element)) {
-        message.author.send("Chill friendo, saying " + element +" isn't allowed!");
-        message.delete().catch(O_o => { });
+    bannedword.forEach(element => {
+        if (message.content.includes(element)) {
+            message.author.send("Chill friendo, saying " + element + " isn't allowed!");
+            message.delete().catch(O_o => { });
+        }
     });
 
     if (command === "!kick") {
